On branch load_reload_using_bowdenlength Changes to be committed: 	modified:   Buttons.cpp 	modified:   motion.cpp 	modified:   permanent_storage.cpp 	modified:   permanent_storage.h
This fix is important for MMU2 units that have configured and calibrated a longer PTFE tube.
Before the fix: unload was using the "stock" length, programmed as a constant + extra length.
This resulted in poor and unnecesarily slow unloading.

The printing time saved, on an average print job with 1000 toolchanges was 2.5 hours!

With this implementation accurate and fast loading and unloading filament using the configured bowden
length of the tubes is realized.

For obtaining the bowden length for the loaded tube, an additional method: BowdenLength::getPrevious()
was implemented/ This ensures the correct unload distance (for the old bowden, rather than the material to be loaded
material) could be obtained.

When checking for all occurrences of the unload routine, also the calibration routine (in Buttons.cpp)
was discovered and had to reflect the correct "previous" material used, or the obtained bowdenlenth would
be incorrect.

Constants were introduced to reflect the speeds / acceleration / deceleration / and points of
acceleration/deceleration, so that they can easily be tuned, independently of changing the program.
These constants are in motion.cpp.

I used the "fast" settings on my Prusa MK3/MMU during 50 hours of printing successfully for ~3000 toolchanges.
They were tested with PLA. The settings as provided contain the somewhat slower load/unload speeds.
These are the values used I found in the code as they were before this change.

It is possible that this is required for e.g. flex material, so I was conservative bumping up the speed
constants to the FAST values that I use.